editing portfolio category is now possible. way to hard to edit photo in place so will leave that out
