+++ src/components/Work/Work.jsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import { useSelector } from 'react-redux';
+import WorkForm from '../WorkForm/WorkForm';
+import CrudBtnsContainer from '../CrudBtnsContainer/CrudBtnsContainer';
+import { portfolioSubmitHandler } from '../../utils/forms/submitHandlers';
+import './Work.scss';
+
+function Work(props) {
+  const { id, category, imageUrl } = props;
+
+  const workForm = useSelector((state) => state.form.WorkForm);
+  const editingForm = useSelector((state) => state.adminDashReducer.editingForm);
+
+  const collection = 'portfolio';
+  const existingData = {
+    category,
+    imageUrl
+  }
+
+  function handleSubmit() {
+    portfolioSubmitHandler(workForm, editingForm);
+  }
+
+  return (
+    <article className="Work" data-testid="Work">
+      {
+        editingForm === id ?
+        <WorkForm
+          initialValues={existingData}
+          onSubmit={handleSubmit}
+        /> :
+        <>
+          <img src={imageUrl} alt="" style={{width: "300px"}}/>
+          <CrudBtnsContainer
+            collection={collection}
+            id={id}
+            category={category}
+          />
+        </>
+      }
+    </article>
+  )
+}
+
+export default Work;
