+++ src/components/Work/Work.scss
